+++ src/ts/UIFactory.ts
@@ -26,6 +26,7 @@ import { SettingsToggleButton } from './components/settings/SettingsToggleButton
 import { FullscreenToggleButton } from './components/buttons/FullscreenToggleButton';
 import { UIContainer } from './components/UIContainer';
 import { BufferingOverlay } from './components/overlays/BufferingOverlay';
+import { ResumeOverlay } from './components/overlays/ResumeOverlay';
 import { PlaybackToggleOverlay } from './components/overlays/PlaybackToggleOverlay';
 import { CastStatusOverlay } from './components/overlays/CastStatusOverlay';
 import { TitleBar } from './components/TitleBar';
@@ -93,7 +94,7 @@ export namespace UIFactory {
           },
         },
         {
-          ui: UIFactory.defaultLayouts.smallScreen(),
+          ui: UIFactory.defaultLayouts.smallScreen(config),
           condition: (context: UIConditionContext) => {
             return !context.isAd && !context.adRequiresUi && context.documentWidth < smallScreenSwitchWidth;
           },
@@ -149,7 +150,7 @@ export namespace UIFactory {
           },
         },
         {
-          ui: UIFactory.defaultLayouts.smallScreen(),
+          ui: UIFactory.defaultLayouts.smallScreen(config),
           condition: (context: UIConditionContext) => {
             return !context.isAd && !context.adRequiresUi;
           },
@@ -283,6 +284,7 @@ export namespace UIFactory {
       });
 
       const conditionalComponents = [config.includeWatermark ? new Watermark() : null].filter(e => e);
+      const resumeComponents = config.enableResumeFromLastPosition !== false ? [new ResumeOverlay()] : [];
 
       return new UIContainer({
         components: [
@@ -294,6 +296,7 @@ export namespace UIFactory {
           new TitleBar(),
           new RecommendationOverlay(),
           ...conditionalComponents,
+          ...resumeComponents,
           new DismissClickOverlay({ target: settingsPanel }),
           settingsPanel,
           new ErrorMessageOverlay(),
@@ -359,7 +362,7 @@ export namespace UIFactory {
       });
     }
 
-    export function smallScreen(): UIContainer {
+    export function smallScreen(config: UIConfig = {}): UIContainer {
       const subtitleOverlay = new SubtitleOverlay();
 
       const settingsPanel = buildDefaultSettingsPanel(subtitleOverlay, -1);
@@ -395,6 +398,8 @@ export namespace UIFactory {
         ],
       });
 
+      const resumeComponents = config.enableResumeFromLastPosition !== false ? [new ResumeOverlay()] : [];
+
       return new UIContainer({
         components: [
           subtitleOverlay,
@@ -403,6 +408,7 @@ export namespace UIFactory {
           // Use the touch overlay on mobile devices and the regular playback toggle overlay on desktop browsers
           BrowserUtils.isMobile ? new TouchControlOverlay() : new PlaybackToggleOverlay(),
           new RecommendationOverlay(),
+          ...resumeComponents,
           controlBar,
           new TitleBar({
             components: [